separate out classes, start using await
